+++ @ecomplus/storefront-app/src/components/js/EcCreditCard.js
@@ -1,10 +1,10 @@
 import { _config, i18n, formatMoney } from '@ecomplus/utils'
+import { $ } from '@ecomplus/storefront-twbs'
 import cardValidator from 'card-validator'
 import loadPaymentClient from './../../lib/load-payment-client'
+import InputDocNumber from '#components/InputDocNumber.vue'
 import InputDate from './../_internal/InputDate.vue'
-import InputDocNumber from './../_internal/InputDocNumber.vue'
 import CleaveInput from 'vue-cleave-component'
-import { SlideYUpTransition } from 'vue2-transitions'
 
 import {
   AboutCvv,
@@ -32,8 +32,7 @@ export default {
   components: {
     InputDate,
     InputDocNumber,
-    CleaveInput,
-    SlideYUpTransition
+    CleaveInput
   },
 
   props: {
@@ -351,5 +350,8 @@ export default {
         break
       }
     }
+    $(function () {
+      $('[data-toggle="popover"]').popover()
+    })
   }
 }