+++ @ecomplus/storefront-app/src/views/Account.vue
@@ -1,13 +1,25 @@
 <template>
   <div id="account">
-    <ec-account
-      :ecomPassport="ecomPassport"
+    <the-account
+      :ecom-passport="ecomPassport"
       :customer.sync="customer"
-      :showOrders.sync="showOrders"
+      :is-orders-list.sync="isOrdersList"
       @login="login"
       @logout="resetAccount"
-      @click:order="viewOrder"
-    />
+    >
+      <template #account-form>
+        <ec-account-form :customer.sync="customer"/>
+      </template>
+      <template #account-addresses>
+        <ec-addresses :customer.sync="customer"/>
+      </template>
+      <template #orders-list>
+        <ec-orders-list
+          :ecom-passport="ecomPassport"
+          @click:order="viewOrder"
+        />
+      </template>
+    </the-account>
   </div>
 </template>