+++ @ecomplus/storefront-components/src/TheAccount.vue
@@ -1,0 +1,3 @@
+<script src="./js/TheAccount.js"></script>
+<template lang="html" src="./html/TheAccount.html"></template>
+<style lang="scss" src="./scss/TheAccount.scss"></style>